+++ src/targeted/targets/action-type-target.mjs
@@ -1,41 +1,23 @@
 import { MODULE_NAME } from '../../consts.mjs';
 import { radioInput } from '../../handlebars-handlers/bonus-inputs/radio-input.mjs';
 import { showChecklist } from '../../handlebars-handlers/targeted/targets/checklist-input.mjs';
+import { isMelee, isNatural, isNaturalSecondary, isPhysical, isRanged, isSpell, isThrown, isWeapon } from '../../util/action-type-helpers.mjs';
 import { listFormat } from '../../util/list-format.mjs';
 import { localize, localizeBonusLabel } from '../../util/localize.mjs';
 import { truthiness } from '../../util/truthiness.mjs';
 import { BaseTarget } from './base-target.mjs';
 
-/** @type {FilterFunc} */
-const isNaturalSecondary = (item, action, actionUse) => {
-    const _isNatural = isNatural(item);
-    const isPrimary = action?.data.naturalAttack.primaryAttack;
-    return _isNatural && !isPrimary;
-}
-/** @type {FilterFunc} */
-const isNatural = (item) => {
-    const isAttack = item instanceof pf1.documents.item.ItemAttackPF;
-    return (isAttack && item.subType === 'natural')
-        || !!item.system.weaponGroups?.value?.includes("natural");
-}
-/** @type {FilterFunc} */
-const isSpell = (item, action) => {
-    const isSpell = item instanceof pf1.documents.item.ItemSpellPF;
-    return isSpell || ['msak', 'rsak', 'spellsave'].includes(action?.data.actionType ?? '');
-}
-
 /** @typedef {keyof typeof filterTypes} FilterType */
-/** @typedef {(item: ItemPF, action?: ItemAction, actionUse?: ActionUse | null) => boolean} FilterFunc */
 
 const filterTypes = /** @type {const} */ ({
-    ['is-melee']: { label: '', filter: /**@type {FilterFunc}*/ (_item, action) => ['mwak', 'msak', 'mcman'].includes(action?.data.actionType ?? '') },
+    ['is-melee']: { label: '', filter: isMelee },
     ['is-natural']: { label: '', filter: isNatural },
     ['is-natural-secondary']: { label: '', filter: isNaturalSecondary },
-    ['is-physical']: { label: '', filter: /**@type {FilterFunc}*/ (item, action) => !!item?.isPhysical },
-    ['is-ranged']: { label: '', filter: /**@type {FilterFunc}*/ (_item, action) => ['rcman', 'rwak', 'rsak', 'twak'].includes(action?.data.actionType ?? '') },
+    ['is-physical']: { label: '', filter: isPhysical },
+    ['is-ranged']: { label: '', filter: isRanged },
     ['is-spell']: { label: '', filter: isSpell },
-    ['is-thrown']: { label: '', filter: /**@type {FilterFunc}*/ (_item, action) => action?.data.actionType === 'twak' },
-    ['is-weapon']: { label: '', filter: /**@type {FilterFunc}*/ (item, action) => ['mwak', 'rwak', 'twak'].includes(action?.data.actionType ?? '') || isNatural(item) },
+    ['is-thrown']: { label: '', filter: isThrown },
+    ['is-weapon']: { label: '', filter: isWeapon },
 });
 
 const all = 'all';